Frozen 0 in additional fee's input box
Problem
=======
When adding additional fees on the split page, the default value would
be 0 and you would not be able to delete the 0.

Solution
========
Don't set the input box's value to the output of the function which
parses the value in the input box.

Testing
========
Tested on localhost and problem has gone away.